refactor(common/CognitoIdTokenSerializer): simplify deserialize
